+++ Changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+-   The ErrorBoundary component was broken if the Redux state wasn't in a
+    supported state.
+
 ## 6.17.0 - 2023-01-18
 
 ### Changed